Refactor(style): revert before component structure
